Minimal C test - printing loop
This is the same program as qira_tests/bin/loop but written as a
statically linked C program that does not use libc, and is compiled for
minimal code size.

This should be akin to the simple hello.asm test for purposes of easily
verifying Qira's behavior (and stuff like QEMU trace output), while
still being cross-platform.