Qemu trace and debug logs comparing qira patches
These are log files I used to compare the qemu traces and qira-related
debug logs to determine if there were any regressions in the binary
trace logs generated, since they were different between the patch for
qemu 2.12 and qemu 3.1.0.

The logs are committed as an archive, because in total they're over 600
MB uncompressed. The archive has the following structure:

The root has directories for each test case run: loop and minloop.
Loop is the standard loop test from `test_auto/source-autogen/loop.c`.
Minloop is the same as loop, but without using anything from libc to
achieve deterministic execution (as much as possible with qemu).

Within each of those is a directory for the platform used, either 16.04
or local. "16.04" is the docker container running on Ubuntu 16.04, and
"local" is when qemu was just on my local machine, running Arch Linux
updated to whatever was current in July of 2019.

Within each platform directory is the directory for the qemu version,
2.5.1 or 3.1.0, and within there are directories for traces from each
run of the binary.

Each trace directory has the following files:
* `0` - the binary trace file
* `0_base` - The contents of `/proc/self/maps` for qemu when the trace
was run. The program's maps are in there, but it clearly also has maps
for things that only qemu uses (or are mapped in by qemu for some
reason).
* `0_env` - Contents of the stack for the binary being traced, just
before the entry point is executed.
* `0_qemu` - Debug log from qemu, including TCG-related logging about
each operation executed, as well as qira-specific logging for each
change that is included in the binary trace file, etc.
* `0_qemu_chng` - Same as above but filtered to only include the logs
lines that correspond to data which is output into the binary trace
file (e.g. setting changelists and recording of changes within each
changelist).
* `0_strace` - strace log for the binary being traced

The traces for the loop test also include `0_qemu_proc` and
`0_qemu_proc_chng`, which are the same logs, but filtered to not include
log lines for instructions executed by library code. This cuts down on
the noise significantly, but for your own sanity if you're reviewing
these files, the minloop test traces are vastly smaller.

There are also a couple of other random files:
* `loop/16.04/qemu-2.5.1/ld-2.23.so` - Cannot recall why this is here
anymore, maybe I needed a specific version of ld inside of the docker
container?
* `filter_proc_changelists.rb` - Ruby script to filter the log files to
do the filtering out of library code mentioned above.